requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.6
django-celery-results==2.0.1
django-rest-framework==0.1.0
djangorestframework==3.12.2
kombu==5.0.2
librabbitmq==2.0.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
